unit tests for extraction and misc refactoring
